CSP-2380 Refactor V1 course provider details API model
